refactor(cost-basis): share receipt integrity primitives
