+++ src/services/portfolio_transaction_processing_service/app/infrastructure/cost_basis/receipt_integrity.py
@@ -5,17 +5,12 @@ from typing import cast
 from portfolio_common.domain.calculation_lineage import (
     CalculationLineage,
     calculation_lineage_from_payload,
-    canonical_content_hash,
+)
+from portfolio_common.domain.cost_basis_receipt_integrity import (
+    receipt_version_content_hash as _receipt_version_content_hash,
 )
 
-
-def required_calculation_lineage(payload: object, context: str) -> CalculationLineage:
-    """Rehydrate required lineage without coupling receipt adapters to JSON storage."""
-
-    lineage = calculation_lineage_from_payload(payload)
-    if lineage is None:
-        raise ValueError(f"{context} is required")
-    return lineage
+__all__ = ["receipt_version_content_hash", "required_calculation_lineage"]
 
 
 def receipt_version_content_hash(
@@ -25,16 +20,23 @@ def receipt_version_content_hash(
     receipt_version: int,
     previous_receipt_content_hash: str | None,
 ) -> str:
-    """Hash one immutable receipt version and its predecessor pointer."""
+    """Expose the shared immutable-version hash with a local typed boundary."""
 
     return cast(
         str,
-        canonical_content_hash(
-            {
-                "previous_receipt_content_hash": previous_receipt_content_hash,
-                "receipt_id": receipt_id,
-                "receipt_version": receipt_version,
-                "semantic_content_hash": semantic_content_hash,
-            }
+        _receipt_version_content_hash(
+            receipt_id=receipt_id,
+            semantic_content_hash=semantic_content_hash,
+            receipt_version=receipt_version,
+            previous_receipt_content_hash=previous_receipt_content_hash,
         ),
     )
+
+
+def required_calculation_lineage(payload: object, context: str) -> CalculationLineage:
+    """Rehydrate required lineage without coupling receipt adapters to JSON storage."""
+
+    lineage = calculation_lineage_from_payload(payload)
+    if lineage is None:
+        raise ValueError(f"{context} is required")
+    return lineage
